% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_as_heading.R
\name{add_as_heading}
\alias{add_as_heading}
\alias{add_as_section}
\alias{add_as_subsection}
\alias{add_as_heading1}
\alias{add_as_heading2}
\alias{add_as_heading3}
\alias{add_as_heading4}
\alias{add_as_heading5}
\alias{add_as_heading6}
\title{Add text as a heading (section name) to \code{knitrContainer}}
\usage{
add_as_heading(container = NULL, obj, level = 1)

add_as_section(container = NULL, obj, level = 1)

add_as_subsection(container = NULL, obj)

add_as_heading1(container = NULL, obj)

add_as_heading2(container = NULL, obj)

add_as_heading3(container = NULL, obj)

add_as_heading4(container = NULL, obj)

add_as_heading5(container = NULL, obj)

add_as_heading6(container = NULL, obj)
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{obj}{(string) a piece of text to be used as heading.}

\item{level}{(integer) The level of heading (between 1 and 6).
1 is the top level section.}
}
\description{
\code{add_as_heading()} converts \code{obj} to \code{\link[base]{character}},
 formats it as a \bold{heading of a section} and includes it in the
 \code{container}. Function \code{print_objects} will print the object as text.
}
\details{
All the remaining functions listed here are wrappers of
  \code{add_as_heading}  with different value of parameter
  \code{level} passed to \code{add_as_heading}:\cr

 In \code{add_as_section()} default value of  \code{level = 1}.\cr
 In \code{add_as_heading1()} default value of  \code{level = 1}.\cr

 In \code{add_as_subsection()} default value of  \code{level = 2}.\cr
 In \code{add_as_heading2()} default value of  \code{level = 2}.\cr

 In \code{add_as_heading3()} default value of  \code{level = 3}.\cr

 In \code{add_as_heading4()} default value of  \code{level = 4}.\cr

 In \code{add_as_heading5()} default value of  \code{level = 5}.\cr

 In \code{add_as_heading6()} default value of  \code{level = 6}.\cr
}
\examples{
# Find more examples in link `knitrContainer-class`


# Initialize container
container <- knitrContainer()

# Fill the container

container \%<>\% add_as_heading1(obj = "Heading 1")
container \%<>\% add_as_heading2(obj = "Heading 2")
container \%<>\% add_as_heading3(obj = "Heading 3")
container \%<>\% add_as_heading4(obj = "Heading 4")
container \%<>\% add_as_heading5(obj = "Heading 5")
container \%<>\% add_as_heading6(obj = "Heading 6")

container \%<>\% add_as_heading(obj = "Heading 5", level = 5)
container \%<>\% add_as_heading(obj = "Heading 6", level = 6)

container \%<>\% add_as_section(obj = "Heading 1")
container \%<>\% add_as_subsection(obj = "Heading 2")

# Print the contents of the container
print_all(container)


# Operator `\%<>\%` updates the object `container`
  ?`\%<>\%`

# These 2 lines are equivalent:
 container  <-  add_as_section(container, obj = "Heading 1")
 container \%<>\% add_as_section(obj = "Heading 1")


}
\seealso{
Other \code{knitrContainer} functions: \code{\link{Join}},
  \code{\link{add_as_}}, \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}},
  \code{\link{print_all}}, \code{\link{print}},
  \code{\link{summary}}
}
\author{
Vilmantas Gegzna
}
