% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_as_section.R
\name{add_as_section}
\alias{add_as_heading1}
\alias{add_as_heading2}
\alias{add_as_heading3}
\alias{add_as_heading4}
\alias{add_as_section}
\alias{add_as_subsection}
\title{Add object as a heading/section name to \code{knitrContainer}}
\usage{
add_as_section(container = NULL, obj, level = 1)

add_as_subsection(container = NULL, obj = "")

add_as_heading1(container = NULL, obj = "")

add_as_heading2(container = NULL, obj = "")

add_as_heading3(container = NULL, obj = "")

add_as_heading4(container = NULL, obj = "")
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{obj}{An object to be added to the \code{container}.}

\item{level}{The level of heading/section to be added. Default is 1
(top level section).}
}
\description{
\code{add_as_section()} converts \code{obj} to \code{\link[base]{character}},
 formats it as a \bold{heading of a section} and includes it in the
 \code{container}. Function \code{print_objects} will print the object as text.
}
\details{
All the remaining functions listed here are wrappers of
  \code{add_as_section}  with different value of parameter
  \code{level} passed to \code{add_as_section}:\cr

 In \code{add_as_heading1()} the default value of  \code{level = 1}.\cr

 In \code{add_as_subsection()} the default value of  \code{level = 2}.\cr
 In \code{add_as_heading2()}  the default value of  \code{level = 2}.\cr

 In \code{add_as_heading3()} the default value of  \code{level = 3}.\cr

 In \code{add_as_heading4()} the default value of  \code{level = 4}.\cr
}
\examples{
# Find more examples in link `knitrContainer-class`


# Initialize container
container <- knitrContainer()

# Fill the container

container \%<>\% add_as_section(obj = "Heading 1")
container \%<>\% add_as_subsection(obj = "Heading 2")

container \%<>\% add_as_heading1(obj = "Heading 1")
container \%<>\% add_as_heading2(obj = "Heading 2")
container \%<>\% add_as_heading3(obj = "Heading 3")
container \%<>\% add_as_heading4(obj = "Heading 4")

container \%<>\% add_as_section(obj = "Heading 5", level = 5)
container \%<>\% add_as_section(obj = "Heading 6", level = 6)

# Extract and print the contents of the container
print_all(container)


# Operator `\%<>\%`
  ?`\%<>\%`

# These 2 lines are equivalent:
 container  <-  add_as_section(container, obj = "Heading 1")
 container \%<>\% add_as_section(obj = "Heading 1")


}
\author{
Vilmantas Gegzna
}
\seealso{
Other X.code.knitrContainer..functions: \code{\link{Join}},
  \code{\link{add_as_}}, \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}},
  \code{\link{print_all}}, \code{\link{print}},
  \code{\link{summary}}
}

