% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/added_as.R
\name{added_as}
\alias{added_as}
\title{[!!!] Set or get attribute `added_as`}
\usage{
added_as(obj, TYPE = NULL, force.TYPE = FALSE)
}
\arguments{
\item{obj}{An object to be updated}

\item{TYPE}{string to be added as attribute \code{$added_as} if
\code{$added_as} is not already present.}

\item{force.TYPE}{Logical. If \code{TRUE}, value of attribute \code{$added_as}
is added or replaced even if \code{obj} already has it.}
}
\description{
Function either returns value of `attributes(obj)$added_as` (if TYPE is not provided)
OR
returns updated `obj` with addedd attribute if both  conditions are fulfilled:
    `TYPE` is provided and attribute is `NULL`. otherwise returns `obj` intact. \cr
}
\details{
If force.TYPE == TRUE, new value of `TYPE` will be added any way.
}
\examples{

# [NO EXAMPLES YET]

}
\author{
Vilmantas Gegzna
}

