% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_obj.R
\name{attach_obj}
\alias{attach_obj}
\title{Attach an object to the end of a list}
\usage{
attach_obj(x = list(), obj)
}
\arguments{
\item{x}{A list.}

\item{obj}{An object to be attached as the last element of the
list \code{x}.}
}
\value{
The list \code{x} with \code{obj} attached as its
        last and unnamed element.
}
\description{
Attach an object to the end of a list
}
\examples{

l <- list(a = 1, b = 2)
attach_obj(l, obj = 3)
 #> $a
 #> [1] 1
 #>
 #> $b
 #> [1] 2
 #>
 #> [[3]]
 #> [1] 3


}

