% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_thumbnail.R
\name{create_thumbnail}
\alias{create_thumbnail}
\title{Create a thumbnail of an image.}
\usage{
create_thumbnail(figPath, tnPath = sub("(.*)(\\\\..*?)$", "\\\\1_small\\\\2",
  figPath), w = 300, h = NULL, ratio = NULL, quality = 85,
  overwrite = FALSE, ...)
}
\arguments{
\item{figPath}{(vector of stings) relative or full path to original figure.}

\item{tnPath}{(vector of stings) relative or full path to thumblail.
By default, suffix "_small" is added to file name.}

\item{w}{(integer|\code{NULL}) width of resized image in pixels.  Default is \code{300}.}

\item{h}{(integer|\code{NULL}) height of resized image in pixels. Default is \code{NULL}.}

\item{ratio}{(numeric|\code{NULL}) expected aspect ratio (width/height).
If \code{NULL} (default), aspepct ratio of the original image is used.}

\item{quality}{(numeric) a numeric ranging from 1 to 100 (default)
controlling the quality of the JPEG output.}

\item{overwrite}{(FALSE|TRUE) logical value, that controls if an exsisting
thumbnail with the same name should be overwritten. Default is \code{FALSE}.}

\item{...}{further parameters (except x, files and quality) to be passed to
function \code{\link[EBImage]{writeImage}()}.}
}
\value{
Function creates a thumbnail and returns a path to to it.
}
\description{
Create a thumbnail of an image.
}
\examples{
\dontrun{
\dontCheck{
figPath = "Figures/2016-08-02-11875.jpg"

# Full name of a thumbnail
sep     <- .Platform$file.sep
tnPath  <- sub(pattern     = paste0("(.*", sep, ")?(Figures)(", sep, ".*)"),
              replacement = paste0("\\\\1\\\\2_small\\\\3"),
             x = figPath)

# Creates a thumbnail and returns a path to it.

tn <- create_thumbnail(figPath)

browseURL(tn)
file.remove(tn)

}}

}
