% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_and_print.R
\name{extract_and_print}
\alias{extract_and_print}
\alias{print_objects}
\title{[!] Print objects from \code{knitrContainer}}
\usage{
extract_and_print(container, ...)

print_objects(container, ...)
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{...}{not used.}
}
\description{
Function is designed to be used in a \code{knitr} file which generates
 \code{HTML} output.\cr
 The function takes every element (i.e., object) of a
 \code{knitrContainer}, prints it appropriately:
 either as text (using \code{\link[base]{cat}}),
 as htmlwidget (using \code{\link[base]{cat}}) or object as-is using
 \code{\link[base]{print}}. Then attaches \code{HTML} dependencies for
 \code{\link[htmlwidgets]{htmlwidgets}} to show them correctly.
}
\details{
Functions \code{extract_and_print()} and \code{print_objects()} are aliases.
}
\examples{

# Find examples in link `knitrContainer-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other X.code.knitrContainer..functions: \code{\link{Join}},
  \code{\link{add_as_section}}, \code{\link{add_as_}},
  \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}}, \code{\link{print}},
  \code{\link{summary}}
}

