% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{[.knitrContainer}
\alias{$.knitrContainer}
\alias{[.knitrContainer}
\alias{[[.knitrContainer}
\title{Extract and Replace parts of knitrContainer objects}
\usage{
\method{"["}{knitrContainer}(x, i, ..., section = NULL, level = 1,
  drop = FALSE)

\method{"[["}{knitrContainer}(x, i, ..., drop = TRUE)

\method{"$"}{knitrContainer}(x, name)
}
\arguments{
\item{x}{a \code{knitrContainer} Object}

\item{i}{index of object inside knitrContainer}

\item{...}{ignored}

\item{drop}{For \code{[[}: drop unnecessary dimensions, see
\code{\link[base]{drop}} and \code{\link[base]{Extract}}. Ignored for
\code{[}, as otherwise invalid \code{knitrContainer} objects might result.}

\item{name}{name of the data column to extract.}

\item{j}{selecting ...}

\item{l}{selecting ...}

\item{name}{name of the data column to extract.}
}
\description{
The Methods to extract and replace elements of a \code{knitrContainer} object
}
\details{
These Methods can be used for selecting/deleting,
and extracting or setting elements of a \code{knitrContainer} object.

\emph{Extracting: \code{[}, \code{[[}, and \code{$}}.

The version with single square brackets (\code{[}) returns the
resulting \code{knitrContainer} object.

\code{[[} yields corresponding \code{knitrContainer}
element returned.

\code{$} returns the selected element.

\emph{Shortcuts.} Three shortcuts to conveniently extract much
 needed parts of the object are defined:

\code{x[[]]} is equivalent to \code{print_all}.

\code{x$.} returns ...

\code{x$..} returns ...

\emph{Replacing: \code{[<-}, \code{[[<-}, and \code{$<-}}.

\code{value} gives the values to be assigned.\cr

\code{[[<-} replaces parts of the ...

\code{[<-} replaces parts of the ...

\code{$<-} replaces a ...

\code{$..<-} is again an abbreviation for the ...
}
\keyword{manip}
\keyword{methods}

