% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join.R
\name{Join}
\alias{Join}
\title{[!] Merge objects to `knitrContainer`}
\usage{
Join(..., list = NULL)
}
\arguments{
\item{...}{objects to be merged.}

\item{list}{objects to be merged provided as a list.}
}
\value{
knitrContainer
}
\description{
\code{Join} converts objects to \code{knitrContainer} objects
(method "as is") and \bold{merges} them to one container.
}
\examples{

# Find examples in link `knitrContainer-class`

}
\seealso{
Other \code{knitrContainer} functions: \code{\link{add_as_heading}},
  \code{\link{add_as_}}, \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}},
  \code{\link{print_all}}, \code{\link{print}},
  \code{\link{summary}}
}
\author{
Vilmantas Gegzna
}
