% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.knitrContainer.R
\name{print}
\alias{print}
\alias{print.knitrContainer}
\title{Print and summary methods for \code{knitrContainer}}
\usage{
\method{print}{knitrContainer}(x, n = 30, preview = 25, units = "Kb", ...)
}
\arguments{
\item{x}{\code{knitrContainer} object.}

\item{n}{The number of rows (objects) to display. Default \code{n = 30}.}

\item{preview}{length of text (number of characters) to be previewed in summary.
Default is 25.}

\item{units}{the units to be used in formatting and printing the size.
    Allowed values for the different \code{standard}s are
    \describe{
      \item{\code{standard = "legacy"}:}{
	\code{"b"}, \code{"Kb"}, \code{"Mb"}, \code{"Gb"}, \code{"Tb"}, \code{"Pb"},
	\code{"B"}, \code{"KB"}, \code{"MB"}, \code{"GB"}, \code{"TB"} and \code{"PB"}.}
      \item{\code{standard = "IEC"}:}{
	\code{"B"}, \code{"KiB"}, \code{"MiB"}, \code{"GiB"},
	\code{"TiB"}, \code{"PiB"}, \code{"EiB"}, \code{"ZiB"} and \code{"YiB"}.}
      \item{\code{standard = "SI"}:}{
	\code{"B"}, \code{"kB"}, \code{"MB"}, \code{"GB"}, \code{"TB"},
	\code{"PB"}, \code{"EB"}, \code{"ZB"} and \code{"YB"}.}
    }
    For all standards, \code{unit = "auto"} is also allowed.
    If \code{standard = "auto"}, any of the "legacy" and \acronym{IEC}
    units are allowed.
    See \sQuote{Formatting and printing object sizes} for details.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Print summary of \code{knitrContainer} object.
}
\examples{

# Find examples in link `knitrContainer-class`

}
\seealso{
Other \code{knitrContainer} functions: \code{\link{Join}},
  \code{\link{add_as_heading}}, \code{\link{add_as_}},
  \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}},
  \code{\link{print_all}}, \code{\link{summary}}
}
\author{
Vilmantas Gegzna
}
