% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all.R
\name{print_all}
\alias{extract_and_print}
\alias{print_all}
\alias{print_objects}
\title{[!] Print objects from \code{knitrContainer}}
\usage{
print_all(container, env = parent.frame(), widget_as_html = "auto", ...)

extract_and_print(container, env = parent.frame(), ...)

print_objects(container, env = parent.frame(), ...)
}
\arguments{
\item{container}{\code{knitr_container} object.}

\item{env}{Environment in which evaluation of expressions an assignments
(objects added with \code{\link{add_as_cmd}} and \code{\link{add_as_data}})
take place.}

\item{widget_as_html}{(TURE, FALSE, "auto") Should htmlwidget be printed as
an HTML code or as an object (default is \code{"auto"}).}

\item{...}{not used.}
}
\description{
Function is designed to be used in a \code{knitr}/\code{rmarkdown} file
 which generates \code{HTML} output.\cr
 The function takes every element (i.e., object) of a
 \code{knitrContainer}, prints it appropriately:
 either as text (using \code{\link[base]{cat}}),
 as htmlwidget (using \code{\link[base]{cat}}) or object as-is using
 \code{\link[base]{print}}. Then attaches \code{HTML} dependencies for
 \code{\link[htmlwidgets]{htmlwidgets}} to show them correctly.
}
\details{
Functions \code{print_objects()} and \code{extract_and_print()} are deprecated.
}
\examples{

# Find examples in link `knitrContainer-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other X.code.knitrContainer..functions: \code{\link{Join}},
  \code{\link{add_as_heading}}, \code{\link{add_as_}},
  \code{\link{as.knitrContainer}},
  \code{\link{knitrContainer-class}}, \code{\link{print}},
  \code{\link{summary}}
}

