% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.knitrContainer.R
\name{summary}
\alias{summary}
\alias{summary.knitrContainer}
\title{Print and summary methods for \code{knitrContainer}}
\usage{
\method{summary}{knitrContainer}(object, n = 100, preview = 40,
  units = "Kb", ...)
}
\arguments{
\item{object}{\code{knitrContainer} object.}

\item{n}{Either the number of rows (objects) to display or numbers of
the first and the last row \code{c(from, to)} to display.
Default \code{n = 100}.}

\item{preview}{number of characters to be previewed in summary.
Default is 40.}

\item{units}{the units to be used in printing the size.  Allowed
    values are
    \code{"b"}, \code{"Kb"}, \code{"Mb"}, \code{"Gb"}, \code{"Tb"}, \code{"Pb"},
    \code{"B"}, \code{"KB"}, \code{"MB"}, \code{"GB"}, \code{"TB"}, \code{"PB"},
    \code{"KiB"}, \code{"MiB"}, \code{"GiB"}, \code{"TiB"},
    \code{"PiB"}, \code{"EiB"}, \code{"ZiB"}, \code{"YiB"},
    and \code{"auto"} (see \sQuote{Details}).  Can be abbreviated.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Print summary of \code{knitrContainer} object.
}
\examples{

# Find examples in link `knitrContainer-class`

}
\author{
Vilmantas Gegzna
}
\seealso{
Other X.code.knitrContainer..functions: \code{\link{Join}},
  \code{\link{add_as_section}}, \code{\link{add_as_}},
  \code{\link{as.knitrContainer}},
  \code{\link{extract_and_print}},
  \code{\link{knitrContainer-class}}, \code{\link{print}}
}

