% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_knitrContainer-package.R
\docType{package}
\name{knitrContainer-package}
\alias{knitrContainer-package}
\title{Print \pkg{pander}, \pkg{plotly} and other objects from inside a loop in a \code{knitr} report}
\description{
\code{knitrContainer} is an R package designed to collect objects and print
them in a \pkg{knitr} report.\cr
}
\section{Objects}{

\bold{What objects can be collectd and printed with `knitrContainer`?}\cr

 An object used by \code{knitrContainer} can be any object, that is
includable in R list an that is printable, such as:\cr
\pkg{pander} tables
(\href{http://rapporter.github.io/pander/}{pander: An R Pandoc Writer},
\code{\link[pander]{pander}});\cr
\pkg{ggplot2} (\href{http://ggplot2.org/}{ggplot2 is a plotting system},
\code{\link[ggplot2]{ggplot}});\cr
\pkg{plotly} (\href{https://plot.ly/r/}{Plotly R Library},
\code{\link[plotly]{plot_ly}});\cr
 text;\cr
 etc.\cr\cr
}

\section{About documentation}{

[+] - function is well described. \cr

[!] - a description is incomplete and needs revision.\cr\cr
}
\author{
Vilmantas Gegzna
}
\references{
Some ideas for functions of \code{knitrContainer} package are taken from
 \href{https://github.com/ropensci/plotly/issues/273#issuecomment-217982315}{this answer} on github.com.
}
\seealso{
\code{knitrContainer}'s
\href{https://github.com/GegznaV/knitrContainer}{gitHub repository}.
}

