% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_knitrContainer_description.R, R/knitrContainer.R
\docType{package}
\name{knitrContainer}
\alias{as.knitrContainer}
\alias{is.knitrContainer}
\alias{knitrContainer}
\alias{knitrContainer-package}
\title{Print \pkg{pander} and \pkg{plotly} objects and  from inside a loop in a \code{knitr} report.}
\usage{
knitrContainer()

is.knitrContainer(obj)

as.knitrContainer(obj = NULL)
}
\arguments{
\item{obj}{object to be tested or converted to \code{knitrContainer} object.}
}
\description{
\code{knitrContainer} is an R package designed to collect objects and print
them in a \pkg{knitr}report.\cr

\code{knitrContainer()} - creates a new empty \code{knitrContainer} object.

\code{is.knitrContainer()} - tests if class of an object is object
                            \code{knitrContainer()}.

\code{as.knitrContainer()} - converts an object to \code{knitrContainer}.
 More specifically:\cr
for list - only the class attribute is added,\cr
for other objects - firstly an object is inctluded into list by
\code{list(object)} and then class attribute is added;\cr
for \code{knitrContainer} objects - they are returned as-is.
}
\section{Objects}{

\bold{What objects can be collectd and printed with `knitrContainer`?}\cr

 An object used by `knitrContainer` can be any object, that is
includable in R list an that is printable, such as objects,
printable with `pander`, text, as well as `ggplot2`, `plotly`
plots, etc. \cr

Examples of hese objects are
\pkg{pander} tables
(\href{http://rapporter.github.io/pander/}{pander: An R Pandoc Writer},
\code{\link[pander]{pander}}),
\pkg{ggplot2} (\href{http://ggplot2.org/}{ggplot2 is a plotting system},
\code{\link[ggplot2]{ggplot}}),
\pkg{plotly} (\href{https://plot.ly/r/}{Plotly R Library},
\code{\link[plotly]{plot_ly}}),
 text and other objects, that can be collected into list.\cr\cr
}

\section{About documentation}{

[+] - function is well described. \cr

[!] - a description is incomplete and needs revision.\cr\cr
}
\examples{

# Find examples in link `knitrContainer-class`

}
\author{
Vilmantas Gegzna

Vilmantas Gegzna
}
\references{
Some ideas for functions of \code{knitrContainer} package are taken from
 \href{https://github.com/ropensci/plotly/issues/273#issuecomment-217982315}{this answer} on github.com.
}
\seealso{
\code{knitrContainer}'s
\href{https://github.com/GegznaV/knitrContainer}{gitHub repository}.

Other X.code.knitrContainer..functions: \code{\link{Join}},
  \code{\link{add_as_section}}, \code{\link{add_as_}},
  \code{\link{knitrContainer-class}},
  \code{\link{print_objects}}, \code{\link{print}},
  \code{\link{summary}}
}

